backend/requirements.txt
fastapi
uvicorn
gunicorn
rembg
pillow
onnxruntime
python-multipart